panel_streamlit/requirements.txt
streamlit==1.39.0
paho-mqtt==1.6.1
